# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pharmaceutical RAG Chat API",
        "description": "\n\uc758\uc57d\ud488 RAG \ucc57\ubd07 \ubc31\uc5d4\ub4dc API\uc785\ub2c8\ub2e4.\n\n## \uc778\uc99d\n- \ucfe0\ud0a4 \uae30\ubc18 \uc138\uc158(`session` \ucfe0\ud0a4, HTTPOnly).\n- `POST /auth/register` \ub610\ub294 `POST /auth/login` \uc131\uacf5 \uc2dc \uc790\ub3d9\uc73c\ub85c \ucfe0\ud0a4\uac00 \uc124\uc815\ub429\ub2c8\ub2e4.\n- \ubaa8\ub4e0 \ubcf4\ud638\ub41c \uc5d4\ub4dc\ud3ec\uc778\ud2b8\ub294 \uc774 \ucfe0\ud0a4\ub97c \uc790\ub3d9 \ud3ec\ud568\ud569\ub2c8\ub2e4(`credentials: \"include\"`).\n\n## \uad8c\ud55c\n- `user` \u2014 \uc77c\ubc18 \uc0ac\uc6a9\uc790: \ucc44\ud305, \ubcf8\uc778 \ud504\ub85c\ud544 \uc870\ud68c/\uc218\uc815.\n- `admin` \u2014 \uad00\ub9ac\uc790: \uc704 \ubaa8\ub4e0 \uad8c\ud55c + \ud504\ub86c\ud504\ud2b8 \uc0dd\uc131/\ubc84\uc804 \uad00\ub9ac/\ud65c\uc131\ud654.\n\n## \ucc44\ud305 \uc2a4\ud2b8\ub9ac\ubc0d\n- `POST /chat/sessions/{session_id}/messages`\ub294 `text/event-stream` (SSE)\uc73c\ub85c \uc751\ub2f5\ud569\ub2c8\ub2e4.\n- \uc774\ubca4\ud2b8 \uc885\ub958: `token`, `tool`, `done`, `error`.\n",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "\ud68c\uc6d0\uac00\uc785",
                "description": "\uc2e0\uaddc \uc0ac\uc6a9\uc790\ub97c \uc0dd\uc131\ud569\ub2c8\ub2e4. \uc131\uacf5 \uc2dc \uc790\ub3d9\uc73c\ub85c \ub85c\uadf8\uc778\ub418\uc5b4 `session` \ucfe0\ud0a4\uac00 \uc124\uc815\ub429\ub2c8\ub2e4.\n\n- `pii_consent`\ub294 \ubc18\ub4dc\uc2dc `true`\uc5ec\uc57c \ud569\ub2c8\ub2e4.\n- \ud504\ub85c\ud544 \ud544\ub4dc(name, age, gender \ub4f1)\ub294 \uac00\uc785 \uc2dc \uc120\ud0dd \uc785\ub825\ud558\uac70\ub098 \uc774\ud6c4 `PATCH /auth/me`\ub85c \uc218\uc815\ud560 \uc218 \uc788\uc2b5\ub2c8\ub2e4.",
                "operationId": "register_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPublic"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc798\ubabb\ub41c \uc694\uccad \ubcf8\ubb38 \ub610\ub294 \ud30c\ub77c\ubbf8\ud130",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "\ub9ac\uc18c\uc2a4 \ucda9\ub3cc (\uc911\ubcf5 \ub4f1)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "\ub85c\uadf8\uc778",
                "description": "\uc774\uba54\uc77c/\ube44\ubc00\ubc88\ud638\ub85c \ub85c\uadf8\uc778\ud569\ub2c8\ub2e4. \uc131\uacf5 \uc2dc `session` \ucfe0\ud0a4\uac00 \uc124\uc815\ub429\ub2c8\ub2e4.",
                "operationId": "login_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "\ub0b4 \ud504\ub85c\ud544 \uc870\ud68c",
                "description": "\ud604\uc7ac \ub85c\uadf8\uc778\ub41c \uc0ac\uc6a9\uc790\uc758 \ud504\ub85c\ud544 \uc804\uccb4\ub97c \ubc18\ud658\ud569\ub2c8\ub2e4.",
                "operationId": "me_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "auth"
                ],
                "summary": "\ub0b4 \ud504\ub85c\ud544 \uc218\uc815",
                "description": "\ud604\uc7ac \ub85c\uadf8\uc778\ub41c \uc0ac\uc6a9\uc790\uc758 \ud504\ub85c\ud544 \ud544\ub4dc\ub97c \ubd80\ubd84 \uc5c5\ub370\uc774\ud2b8\ud569\ub2c8\ub2e4.\n\ubcf4\ub0b8 \ud544\ub4dc\ub9cc \ubcc0\uacbd\ub418\uace0, \uc0dd\ub7b5\ud558\uac70\ub098 `null`\uc778 \ud544\ub4dc\ub294 \uae30\uc874 \uac12\uc744 \uc720\uc9c0\ud569\ub2c8\ub2e4.",
                "operationId": "update_me_auth_me_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserProfileUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail",
                        "description": "\uc0ac\ub78c\uc774 \uc77d\uc744 \uc218 \uc788\ub294 \uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "examples": [
                            "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse",
                "description": "\ubaa8\ub4e0 4xx/5xx \uc751\ub2f5\uc5d0\uc11c \uc0ac\uc6a9\ud558\ub294 \uacf5\ud1b5 \uc5d0\ub7ec \uc2a4\ud0a4\ub9c8."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "LoginRequest"
            },
...
            "RegisterRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "title": "Password"
                    },
                    "pii_consent": {
                        "type": "boolean",
                        "title": "Pii Consent"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "age": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "symptoms_note": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Symptoms Note"
                    },
                    "current_medications": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Current Medications"
                    },
                    "allergies": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Allergies"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "pii_consent"
                ],
                "title": "RegisterRequest"
            },
...
            "UserProfileUpdate": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "age": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "symptoms_note": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Symptoms Note"
                    },
                    "current_medications": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Current Medications"
                    },
                    "allergies": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Allergies"
                    }
                },
                "type": "object",
                "title": "UserProfileUpdate"
            },
            "UserPublic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "age": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age"
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gender"
                    },
                    "symptoms_note": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Symptoms Note"
                    },
                    "current_medications": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Current Medications"
                    },
                    "allergies": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Allergies"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "role"
                ],
                "title": "UserPublic"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    },
...
}